Update identifier indices for inline examples
SUSHI currently resolves soft indexing in a way that does not account
for existing named slices. As a result, the identifier slices on these
examples were having the values in the "mpid" slice overwritten. This
update sets the indices such that there are two elements in the
identifier arrays. The manual slice ordering feature that will be part
of SUSHI v3.0 will allow soft indexing to be used here, if the feature
is enabled.